+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -159,5 +159,6 @@ public let masterRuleList = RuleList(rules: [
     VoidReturnRule.self,
     WeakDelegateRule.self,
     XCTFailMessageRule.self,
+    XCTSpecificMatcherRule.self,
     YodaConditionRule.self
 ])